--- frontend/src/components/sortableTask/SortableTask.css
@@ -1,62 +1,0 @@
-.task-card {
-  background-color: #ffffff;
-  border-radius: 10px;
-  padding: 18px;
-  border: 1px solid #e8e8e8;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.08);
-  transition: all 0.2s ease;
-  cursor: pointer;
-  text-align: left;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.task-header {
-  display: flex;
-  align-items: flex-start;
-  gap: 10px;
-}
-
-.task-icon {
-  color: #9b9b9b;
-  font-size: 18px;
-  flex-shrink: 0;
-  margin-top: 2px;
-  cursor: grab;
-}
-
-.task-title {
-  font-size: 15px;
-  font-weight: 600;
-  color: #1a1a1a;
-  margin: 0;
-  flex: 1;
-  line-height: 1.4;
-  letter-spacing: -0.1px;
-}
-
-.task-description {
-  font-size: 13px;
-  color: #6b6b6b;
-  line-height: 1.6;
-  margin: 0;
-}
-
-.task-footer {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  margin-top: auto;
-}
-
-.task-status {
-  display: inline-block;
-  padding: 5px 12px;
-  border-radius: 12px;
-  font-size: 11px;
-  font-weight: 600;
-  letter-spacing: 0.2px;
-  text-transform: uppercase;
-  white-space: nowrap;
-}
